# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Food diary API
 * OpenAPI spec version: 0.1.0
 */
export type MonitoredNutrient = typeof MonitoredNutrient[keyof typeof MonitoredNutrient];


export const MonitoredNutrient = {
  potassium: 'potassium',
  sodium: 'sodium',
  cholesterol: 'cholesterol',
  salt: 'salt',
  fiber: 'fiber',
} as const;

export interface NutrientTargets {
  /** @minimum 1 */
  potassium?: number;
  /** @minimum 1 */
  sodium?: number;
  /** @minimum 1 */
  cholesterol?: number;
}

export interface UserPreferences {
  monitoredNutrients: MonitoredNutrient[];
  nutrientTargets?: NutrientTargets;
}

export interface UserPreferencesInput {
  monitoredNutrients: MonitoredNutrient[];
  nutrientTargets?: NutrientTargets;
}

export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Food {
  id: number;
  /** @nullable */
  barcode?: string | null;
  name: string;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  servingSize?: string | null;
  /** @nullable */
  kcalPerServing?: number | null;
  /** @nullable */
  fatPerServing?: number | null;
  /** @nullable */
  saturatedFatPerServing?: number | null;
  /** @nullable */
  carbsPerServing?: number | null;
  /** @nullable */
  sugarsPerServing?: number | null;
  /** @nullable */
  fiberPerServing?: number | null;
  /** @nullable */
  proteinPerServing?: number | null;
  /** @nullable */
  saltPerServing?: number | null;
  /** @nullable */
  sodiumPerServing?: number | null;
  /** @nullable */
  potassiumPerServing?: number | null;
  /** @nullable */
  cholesterolPerServing?: number | null;
  /** @nullable */
  nutriscoreGrade?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface ParseFoodInput {
  /** @minLength 1 */
  query: string;
}

export interface ParsedFood {
  foodName: string;
  quantity: number;
  unit: string;
  calories: number;
  proteinG: number;
  carbsG: number;
  fatG: number;
}

export interface CreateFoodInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  servingSize?: string | null;
  /** @nullable */
  kcalPerServing?: number | null;
  /** @nullable */
  fatPerServing?: number | null;
  /** @nullable */
  saturatedFatPerServing?: number | null;
  /** @nullable */
  carbsPerServing?: number | null;
  /** @nullable */
  sugarsPerServing?: number | null;
  /** @nullable */
  fiberPerServing?: number | null;
  /** @nullable */
  proteinPerServing?: number | null;
  /** @nullable */
  saltPerServing?: number | null;
  /** @nullable */
  sodiumPerServing?: number | null;
  /** @nullable */
  potassiumPerServing?: number | null;
  /** @nullable */
  cholesterolPerServing?: number | null;
}

export interface FoodLookupInput {
  barcode: string;
}

export interface FoodHistoryEntry {
  id: number;
  foodId: number;
  scannedAt: string;
  food: Food;
}

export type FoodHistoryResponsePagination = {
  limit: number;
  offset: number;
  hasMore: boolean;
};

export interface FoodHistoryResponse {
  items: FoodHistoryEntry[];
  pagination: FoodHistoryResponsePagination;
}

export type DiaryEntryMealType = typeof DiaryEntryMealType[keyof typeof DiaryEntryMealType];


export const DiaryEntryMealType = {
  breakfast: 'breakfast',
  lunch: 'lunch',
  dinner: 'dinner',
  snack: 'snack',
} as const;

export interface DiaryEntry {
  id: number;
  date: string;
  mealType: DiaryEntryMealType;
  servings: number;
  food: Food;
  createdAt: string;
}

export type DiaryEntryInputMealType = typeof DiaryEntryInputMealType[keyof typeof DiaryEntryInputMealType];


export const DiaryEntryInputMealType = {
  breakfast: 'breakfast',
  lunch: 'lunch',
  dinner: 'dinner',
  snack: 'snack',
} as const;

export interface DiaryEntryInput {
  foodId: number;
  /** YYYY-MM-DD */
  date: string;
  mealType: DiaryEntryInputMealType;
  servings: number;
}

export interface DailyDiary {
  date: string;
  breakfast: DiaryEntry[];
  lunch: DiaryEntry[];
  dinner: DiaryEntry[];
  snack: DiaryEntry[];
}

export interface DailySummary {
  date: string;
  totalKcal: number;
  totalFat: number;
  totalCarbs: number;
  totalProtein: number;
  entryCount: number;
}

export interface MealSetItem {
  id: number;
  servings: number;
  food: Food;
}

/**
 * @nullable
 */
export type MealSetCategory = typeof MealSetCategory[keyof typeof MealSetCategory] | null;


export const MealSetCategory = {
  breakfast: 'breakfast',
  lunch: 'lunch',
  dinner: 'dinner',
  snack: 'snack',
} as const;

export interface MealSet {
  id: number;
  name: string;
  /** @nullable */
  category?: MealSetCategory;
  createdAt: string;
  items: MealSetItem[];
}

/**
 * @nullable
 */
export type MealSetInputCategory = typeof MealSetInputCategory[keyof typeof MealSetInputCategory] | null;


export const MealSetInputCategory = {
  breakfast: 'breakfast',
  lunch: 'lunch',
  dinner: 'dinner',
  snack: 'snack',
} as const;

export interface MealSetInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  category?: MealSetInputCategory;
}

/**
 * @nullable
 */
export type MealSetUpdateInputCategory = typeof MealSetUpdateInputCategory[keyof typeof MealSetUpdateInputCategory] | null;


export const MealSetUpdateInputCategory = {
  breakfast: 'breakfast',
  lunch: 'lunch',
  dinner: 'dinner',
  snack: 'snack',
} as const;

export interface MealSetUpdateInput {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  category?: MealSetUpdateInputCategory;
}

export interface MealSetItemInput {
  foodId: number;
  servings: number;
}

export type MealSetLogInputMealType = typeof MealSetLogInputMealType[keyof typeof MealSetLogInputMealType];


export const MealSetLogInputMealType = {
  breakfast: 'breakfast',
  lunch: 'lunch',
  dinner: 'dinner',
  snack: 'snack',
} as const;

export interface MealSetLogInput {
  /** YYYY-MM-DD — defaults to today if omitted */
  date?: string;
  mealType: MealSetLogInputMealType;
}

export type SearchFoodsParams = {
q: string;
};

export type GetFoodHistoryParams = {
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
/**
 * @minimum 0
 */
offset?: number;
};

export type GetDiaryParams = {
/**
 * Date in YYYY-MM-DD format
 */
date: string;
};

export type GetDiarySummaryParams = {
date: string;
};