# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Food diary API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /diary:
    get:
      operationId: getDiary
      tags: [diary]
      summary: Get diary entries for a specific date
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            description: Date in YYYY-MM-DD format
      responses:
        "200":
          description: Diary entries grouped by meal type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyDiary"
  /diary/entries:
    post:
      operationId: addDiaryEntry
      tags: [diary]
      summary: Add a food entry to the diary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiaryEntryInput"
      responses:
        "201":
          description: Entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiaryEntry"
  /diary/entries/{id}:
    delete:
      operationId: deleteDiaryEntry
      tags: [diary]
      summary: Remove a diary entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Entry deleted
        "404":
          description: Entry not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /diary/summary:
    get:
      operationId: getDiarySummary
      tags: [diary]
      summary: Get daily nutrition totals for a date
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Daily nutrition summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailySummary"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Food:
      type: object
      properties:
        id:
          type: number
        barcode:
          type: ["string", "null"]
        name:
          type: string
        brand:
          type: ["string", "null"]
        servingSize:
          type: ["string", "null"]
        kcalPerServing:
          type: ["number", "null"]
        fatPerServing:
          type: ["number", "null"]
        saturatedFatPerServing:
          type: ["number", "null"]
        carbsPerServing:
          type: ["number", "null"]
        sugarsPerServing:
          type: ["number", "null"]
        fiberPerServing:
          type: ["number", "null"]
        proteinPerServing:
          type: ["number", "null"]
        saltPerServing:
          type: ["number", "null"]
        sodiumPerServing:
          type: ["number", "null"]
        potassiumPerServing:
          type: ["number", "null"]
        cholesterolPerServing:
          type: ["number", "null"]
        nutriscoreGrade:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
      required:
        - id
        - name
...
    DiaryEntry:
      type: object
      properties:
        id:
          type: number
        date:
          type: string
        mealType:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        servings:
          type: number
        food:
          $ref: "#/components/schemas/Food"
        createdAt:
          type: string
      required:
        - id
        - date
        - mealType
        - servings
        - food
        - createdAt
    DiaryEntryInput:
      type: object
      properties:
        foodId:
          type: number
        date:
          type: string
          description: YYYY-MM-DD
        mealType:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        servings:
          type: number
      required:
        - foodId
        - date
        - mealType
        - servings
    DailyDiary:
      type: object
      properties:
        date:
          type: string
        breakfast:
          type: array
          items:
            $ref: "#/components/schemas/DiaryEntry"
        lunch:
          type: array
          items:
            $ref: "#/components/schemas/DiaryEntry"
        dinner:
          type: array
          items:
            $ref: "#/components/schemas/DiaryEntry"
        snack:
          type: array
          items:
            $ref: "#/components/schemas/DiaryEntry"
      required:
        - date
        - breakfast
        - lunch
        - dinner
        - snack
    DailySummary:
      type: object
      properties:
        date:
          type: string
        totalKcal:
          type: number
        totalFat:
          type: number
        totalCarbs:
          type: number
        totalProtein:
          type: number
        entryCount:
          type: number
      required:
        - date
        - totalKcal
        - totalFat
        - totalCarbs
        - totalProtein
        - entryCount
...