# lib/api-zod/src/generated/types/dailySummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Food diary API
 * OpenAPI spec version: 0.1.0
 */

export interface DailySummary {
  date: string;
  totalKcal: number;
  totalFat: number;
  totalCarbs: number;
  totalProtein: number;
  entryCount: number;
}