# lib/api-zod/src/generated/types/diaryEntryInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Food diary API
 * OpenAPI spec version: 0.1.0
 */
import type { DiaryEntryInputMealType } from './diaryEntryInputMealType';

export interface DiaryEntryInput {
  foodId: number;
  /** YYYY-MM-DD */
  date: string;
  mealType: DiaryEntryInputMealType;
  servings: number;
}